Implemented the addEdge and hasEdge methods.
I implemented the addEdge method. In doing so I changed the edges variable
to a List from a HashMap. In order to to allow multiple edges to and from
the same nodes the data structure has to be a list and not a HashMap.

I implement the hasEdge method. The method checks the to see if the parameter
matches any of the ID of other edges.